Rydde opp EtteroppgjoerService og EtteroppgjoerForbehandlingService
- Flytt EtteroppgjoerSvarfrist og FantIkkeEtteroppgjoer til EtteroppgjoerModel.kt
  der de naturlig hører hjemme, fjerner unatural import fra child til parent pakke
- Gjør EtteroppgjoerService constructor-params private
- Gjør kanOppretteForbehandlingForEtteroppgjoer private (kun intern bruk)
- Endre validerOppgaveForOpprettForbehandling til å returnere Unit (ren validering)
